requirements.txt
django-registration-redux==1.3
